Refactored setDisplayedChatId
setDisplayedChatId now only selects chat's with messages. This logic correctly moves the pointer for cases where a newChat is made and no messages are sent. Fixes a bug where the user could not select the previously displayed chat after making a new chat.

Additionally, classNames are added for css styling,  such as equal border's for the user-chats-container and correct highlighting of the displayedChat.