Logic for if user can edit
Implemented admin privileges so that user's can only edit their own profile and group's where they are an admin. Admin is automatically granted to the group creator

Additionally set user information (username, bio, photo) to the respective variables so that if the user chooses to edit that variable, the value is initially the current value.